Cluster.canonicalise / reverseEdges helpers + envelope/EOF read fixes
Adds the read-side API for querying the inverse direction of edges that
the v3 write path strips (`contains` / `buildsTo`) and for resolving
alternate identifiers via the aliasMap:

  - `cluster.canonicalise(id)` — merges the per-DataFile aliasMaps into
    one alias → canonical lookup and resolves any identifier to its
    canonical form.
  - `cluster.reverseEdges(id, edgeType)` — lazily builds a
    `(forwardEdgeType, target) → sources` index by streaming every Item
    in every DataFile via `GRDWalker`, then answers reverse-direction
    queries by mapping the requested type to its forward counterpart.
    Handles `aliasTo` specially since alternate-form Items don't exist
    on disk post-collapse.

Fixes two pre-existing bugs uncovered by the new tests, both in code
paths the original test suite never exercised:

  - `IndexFile.open` / `DataFile.open` / `GoatRodeoCluster.open` read
    envelope lengths via `Helpers.readLenAndCBOR`, which uses
    `readInt` (4 bytes), but `writeABlock` / `writeEntries` write the
    length via `writeShort` (2 bytes). The three call sites now use
    `readShort` + `readCBOR` directly.
  - `Helpers.readCBOR` was passing a `ByteBuffer` to the CBOR decoder
    without rewinding the position past the bytes it had just read,
    so the decoder saw end-of-input. Reset the position to 0 before
    decoding.
  - `writeABlock` wrote the end-of-file sentinel via
    `writeShort(-1)` (2 bytes), but `readNext` reads the frame
    length via `readInt` (4 bytes), so the marker was never matched
    and the reader would attempt to allocate a buffer for a garbage
    length. Write the sentinel via `writeInt(-1)` so the formats line
    up. The pre-existing `GraphBenchmark` comment that flags this as
    a known edge case is no longer true.

All 1322 tests pass (1320 prior + 2 new helper tests).

+++ src/main/scala/io/spicelabs/goatrodeo/omnibor/GraphManager.scala
@@ -137,7 +137,11 @@ object GraphManager {
       }
     }
 
-    Helpers.writeShort(writer, -1) // a marker that says end of file
+    // 4-byte -1 sentinel — matches the 4-byte length prefix `readNext`
+    // expects for each item frame, so the reader can detect end-of-file
+    // by reading -1 here. (Older revisions of this file wrote a 2-byte
+    // short, which `readNext`'s `Helpers.readInt` could not recognise.)
+    Helpers.writeInt(writer, -1)
 
     // write final back-pointer (to the last entry record)
     Helpers.writeLong(writer, previousPosition)
